Fix glance tempest plugin test to use the right tempest variables
A set of changes moved the glance specific bits to use the tempest
configuration option for detecting if the secure RBAC tests should be
run [0][1]. But, we didn't update the tests themselves to use the new option,
causing the tests to be skipped without our knowledge.

This commit updates the tests to use the new option and gets them
running again.

[0] https://review.opendev.org/c/openstack/glance/+/778952
[1] https://review.opendev.org/c/openstack/devstack/+/778945

Change-Id: I01e0db7f0d79f072211ee63ea48a6558d008d285